+++ tests/conftest.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 from typing import NoReturn
 import os
 
+from click.testing import CliRunner
 import pytest
 
 if os.getenv('_PYTEST_RAISE', '0') != '0':  # pragma no cover
@@ -16,3 +17,8 @@ if os.getenv('_PYTEST_RAISE', '0') != '0':  # pragma no cover
     @pytest.hookimpl(tryfirst=True)
     def pytest_internalerror(excinfo: pytest.ExceptionInfo[BaseException]) -> NoReturn:
         raise excinfo.value
+
+
+@pytest.fixture
+def runner() -> CliRunner:
+    return CliRunner()